+++ config/environments/staging.rb
@@ -38,6 +38,7 @@ Chill::Application.configure do
 
   # See everything in the log (default is :info)
   # config.log_level = :debug
+  config.log_level = (ENV['RAILS_LOG_LEVEL'] || :debug).to_sym
 
   # Prepend all log lines with the following tags
   # config.log_tags = [ :subdomain, :uuid ]