+++ spec/controllers/entities/search_controller_spec.rb
@@ -2,9 +2,9 @@ require 'rails_helper'
 
 describe Entities::SearchController, type: :controller do
   context '#index' do
-    it 'uses SearchesModels' do
-      searcher = SearchesModels.new
-      allow_any_instance_of(SearchesModels).to receive(:new).and_return(searcher)
+    it 'uses ElasticsearchQuery' do
+      searcher = ElasticsearchQuery.new
+      allow_any_instance_of(ElasticsearchQuery).to receive(:new).and_return(searcher)
 
       get :index, params: { term: 'foo', format: :json }
     end