+++ spec/controllers/notices/search_controller_spec.rb
@@ -2,9 +2,9 @@ require 'rails_helper'
 
 describe Notices::SearchController do
   context "#index" do
-    it "uses SearchesModels" do
-      searcher = SearchesModels.new
-      expect(SearchesModels).to receive(:new).and_return(searcher)
+    it "uses ElasticsearchQuery" do
+      searcher = ElasticsearchQuery.new
+      expect(ElasticsearchQuery).to receive(:new).and_return(searcher)
 
       get :index, params: { term: 'foo' }